mtserver/requirement.txt
Django==2.2.3
django-cors-headers==3.1.1
django-filter==2.2.0
django-shortuuidfield==0.1.3
djangorestframework==3.10.1
mysqlclient==1.4.2
pycryptodomex==3.7.2
Pygments==2.4.2
PyJWT==1.7.1
PyMySQL==0.9.3
python-alipay-sdk==1.10.1
python-memcached==1.59
pytz==2019.1
shortuuid==0.5.0
six==1.12.0
sqlparse==0.3.0